third_party/blink/web_tests/fast/loader/local-svg-parsed-as-svg-expected.txt

ALERT: PASS: File parsed as SVG
